+++ integration-tests/jupyterlab_config.py
@@ -1,6 +1,5 @@
 from IPython.html.services.contents.filemanager import FileContentsManager
 
 c.ServerApp.jpserver_extensions = {"jupyterlab_hubshare": True}
-c.HubShare.share_url_template = {
-    "qs": {"share-path": "{path}"},
-}
+c.HubShare.file_path_template = "{path}"
+c.HubShare.use_jupyterhub_redirect = False