+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# v0.2.0
+- Use b64encoded path otherwise jupyterhub user-redirect will try to encode string again and cause issue.
+- simplify the share-url configuration
 # v0.1.0
 - support share and preview functions